added 'partially working' group transfer
