+++ src/bots/telegram/main.py
@@ -32,10 +32,10 @@ from bots.telegram.group_bill_split import (
     cancel_bill_split_handler,
     confirm_bill_split_callback_handler,
 )
-from bots.telegram.test_group_transfer import (
-    group_transfer_entry,
+from bots.telegram.group_transfer import (
+    cancel_transfer_handler,
     confirm_transfer_handler,
-    cancel_transfer_handler
+    group_transfer_entry,
 )
 from bots.telegram.receipts import receipts_handlers  # New import
 from bots.telegram.reply_handlers import reply_handler
@@ -150,7 +150,9 @@ def main() -> None:
         MessageHandler(filters.ChatType.GROUP, group_chat_handler)
     )
 
-    application.add_handler(MessageHandler(filters.ChatType.GROUP, group_chat_handler))
+    application.add_handler(
+        MessageHandler(filters.ChatType.GROUP, group_chat_handler)
+    )
     application.add_handler(
         CallbackQueryHandler(
             confirm_bill_split_callback_handler, pattern="^confirm_bill_split_"
